tsre-service/requirements.txt
flask==3.0.2
python-json-logger==2.0.7
click==8.1.7
pyyaml==6.0.1